+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -16,6 +16,7 @@ CHANGELOG
  * Added tag `routing.expression_language_function` to define functions available in route conditions
  * Added `debug:container --deprecations` option to see compile-time deprecations.
  * Made `BrowserKitAssertionsTrait` report the original error message in case of a failure
+ * Added ability for `config:dump-reference` and `debug:config` to dump and debug kernel container extension configuration.
 
 5.0.0
 -----